main

+++ src/main.py
@@ -5,20 +5,29 @@ import threading
 import datetime
 import argparse
 import asyncio
+import ctypes
 import sys
 import os
 
 from pathlib import Path
 
-from src.modules import Logger, search_file_in_dirs, ConfigManage, ProcessManager, AccountSwitcher, TASConfigException, \
-    ProcessMonitor, AESCipher
+from src.modules import (
+    search_file_in_dirs,
+    TASConfigException,
+    AccountSwitcher,
+    ProcessManager,
+    ProcessMonitor,
+    ConfigManage,
+    AESCipher,
+    Logger,
+)
 from src.ui import open_help_window, open_settings_window
 
 logger = Logger()
-TITLE = 'TAS'
-VERSION = '1.2.0'
-
+TITLE = "TAS"
+VERSION = "1.2.1"
 CONFIG = ConfigManage()
+kernel32 = ctypes.WinDLL("kernel32")
 
 
 def handle_global_exception(exc_type, exc_value, exc_traceback):
@@ -28,24 +37,36 @@ def handle_global_exception(exc_type, exc_value, exc_traceback):
     logger.exception(
         f"捕获到未处理异常, 请尝试重启程序,\n若问题依旧请联系开发者或发布Issues，开发者会尽快解决该问题.",
         exc_value,
-        popup=True
+        popup=True,
     )
 
 
 def parse_arguments() -> argparse.Namespace:
     """参数解析"""
-    parser = argparse.ArgumentParser(description='参数解析器', add_help=False, exit_on_error=False)
+    parser = argparse.ArgumentParser(
+        description="参数解析器", add_help=False, exit_on_error=False
+    )
 
     action_group = parser.add_argument_group()
-    action_group.add_argument('--encrypt', '-e', action='store_true', help='立即加密')
-    action_group.add_argument('--decrypt', '-d', action='store_true', help='立即解密')
-    action_group.add_argument('--switch', '-s', type=str, metavar='tag', help='切换指定标签的账号')
+    action_group.add_argument("--encrypt", "-e", action="store_true", help="立即加密")
+    action_group.add_argument("--decrypt", "-d", action="store_true", help="立即解密")
+    action_group.add_argument(
+        "--switch", "-s", type=str, metavar="tag", help="切换指定标签的账号"
+    )
 
     exclusive_group = parser.add_mutually_exclusive_group()
-    exclusive_group.add_argument('--version', '-v', action='store_true', help='获取当前版本')
-    exclusive_group.add_argument('--settings', '-c', action='store_true', help='打开设置窗口')
-    exclusive_group.add_argument('--help', '-h', action='store_true', help='获取帮助文档')
-    parser.add_argument('--password', '-p', type=str, metavar='password', help='指定解密密钥')
+    exclusive_group.add_argument(
+        "--version", "-v", action="store_true", help="获取当前版本"
+    )
+    exclusive_group.add_argument(
+        "--settings", "-c", action="store_true", help="打开设置窗口"
+    )
+    exclusive_group.add_argument(
+        "--help", "-h", action="store_true", help="获取帮助文档"
+    )
+    parser.add_argument(
+        "--password", "-p", type=str, metavar="password", help="指定解密密钥"
+    )
     return parser.parse_args()
 
 
@@ -60,13 +81,13 @@ def check_argument() -> str:
     if args.help:
         open_help_window(VERSION)
     elif args.version:
-        logger.info(f'{TITLE} v{VERSION}', popup=True)
+        logger.info(f"{TITLE} v{VERSION}", popup=True)
     elif args.settings:
         open_settings_window(VERSION)
     elif args.encrypt:
-        process_tags('encrypt')
+        process_tags("encrypt")
     elif args.decrypt:
-        process_tags('decrypt')
+        process_tags("decrypt")
     elif args.switch:
         return validate_tag(args.switch)
     else:
@@ -77,36 +98,36 @@ def check_argument() -> str:
 def process_tags(operation: str) -> None:
     """处理所有标签的加密/解密操作"""
     if not CONFIG.pwd:
-        logger.error('未指定密钥.', popup=True)
+        logger.error("未指定密钥.", popup=True)
         sys.exit()
     cipher = AESCipher(CONFIG.pwd)
 
     operation_name = {
-        'encrypt': '加密',
-        'decrypt': '解密',
+        "encrypt": "加密",
+        "decrypt": "解密",
     }.get(operation)
 
     processed_tags = []
 
     for tag in CONFIG.tags:
-        path = Path(CONFIG.path) / search_file_in_dirs(CONFIG.path, tag) / 'key_datas'
+        path = Path(CONFIG.path) / search_file_in_dirs(CONFIG.path, tag) / "key_datas"
         try:
-            if operation == 'encrypt':
+            if operation == "encrypt":
                 cipher.decrypt(path, save=False)
             else:
                 cipher.decrypt(path)
                 processed_tags.append(tag)
         except Exception:
-            if operation == 'encrypt':
+            if operation == "encrypt":
                 cipher.encrypt(path)
                 processed_tags.append(tag)
             else:
-                logger.error(f'解密失败, 密码错误.', popup=True)
+                logger.error(f"解密失败, 密码错误.", popup=True)
 
     if not processed_tags:
-        msg = f'所有标签均已{operation_name}'
+        msg = f"所有标签均已{operation_name}"
     else:
-        msg = f'本次{operation_name}的标签 -> {processed_tags}.'
+        msg = f"本次{operation_name}的标签 -> {processed_tags}."
 
     logger.info(msg, popup=True)
 
@@ -135,20 +156,22 @@ def check_configs() -> bool:
         default_tdata = CONFIG.default
 
         if not os.path.isfile(os.path.join(path, client)):
-            raise TASConfigException('无法找到客户端')
+            raise TASConfigException("无法找到客户端")
 
         if not os.path.isdir(path):
-            raise TASConfigException('路径格式不正确')
+            raise TASConfigException("路径格式不正确")
 
         if not default_tdata:
-            raise TASConfigException('默认的账户未设置')
+            raise TASConfigException("默认的账户未设置")
 
         if not search_file_in_dirs(path, default_tdata):
-            raise TASConfigException(f"默认账户配置无效, 标记为'{default_tdata}'的账户文件夹未找到")
+            raise TASConfigException(
+                f"默认账户配置无效, 标记为'{default_tdata}'的账户文件夹未找到"
+            )
 
         return True
     except TASConfigException as e:
-        logger.error(f'配置验证失败, {e.message}.', popup=True)
+        logger.error(f"配置验证失败, {e.message}.", popup=True)
         return False
     except Exception as e:
         raise e
@@ -184,8 +207,9 @@ def initialize() -> bool:
     sys.excepthook = handle_global_exception
     config_file = CONFIG.config_file
     if not os.path.exists(config_file):
-        logger.error(f'配置文件 {config_file} 不存在.')
+        logger.error(f"配置文件 {config_file} 不存在.")
         open_settings_window(VERSION)
+        sys.exit()
     else:
         try:
             if not check_configs():
@@ -193,43 +217,35 @@ def initialize() -> bool:
                 sys.exit()
         except Exception as e:
             logger.exception(
-                '客户端初始化失败, 请重试.\n若问题依旧请联系开发者或发布Issues，开发者会尽快解决该问题.',
-                e, popup=True
+                "客户端初始化失败, 请重试.\n若问题依旧请联系开发者或发布Issues.",
+                e,
+                popup=True,
             )
             sys.exit()
-    logger.info('初始化成功.')
+    logger.info("初始化成功.")
     return True
 
 
 def main():
     """主函数"""
     try:
-        initialize()
+        if not initialize():
+            return 1
         ProcessManager.kill_process(CONFIG.client)
 
-        loop = asyncio.new_event_loop()
-        monitor = ProcessMonitor(CONFIG.client)
-        watch_thread = threading.Thread(
-            target=run_async_in_thread,
-            args=(loop, watcher(monitor)),
-            daemon=True
-        )
-        watch_thread.start()
-        logger.info('监控线程启动成功.')
+        if CONFIG.tag:
+            loop = asyncio.new_event_loop()
+            monitor = ProcessMonitor(CONFIG.client)
+            watch_thread = threading.Thread(target=run_async_in_thread, args=(loop, watcher(monitor)), daemon=True)
+            watch_thread.start()
+            logger.info("监控线程启动成功.")
 
         AccountSwitcher().process()
         CONFIG.complete = True
 
-        if monitor:
-            logger.info("正在停止监控线程...")
-            watch_thread.join(timeout=5.0)
-            if watch_thread.is_alive():
-                logger.warning("监控线程未正常退出, 强制终止.")
-
-        logger.info('任务完成.')
     finally:
         if not CONFIG.log_output:
-            return
-        with open(os.path.join(os.getcwd(), 'TAS.log'), 'a', encoding='utf-8') as f:
-            f.write(f'{"-" * 20}{datetime.datetime.now()}{"-" * 20}')
-        return
+            return 0
+        with open(os.path.join(os.getcwd(), "TAS.log"), "a", encoding="utf-8") as f:
+            f.write(f"{'-' * 20}{datetime.datetime.now()}{'-' * 20}")
+    return 0
